Finished splitting up the writer.
