+++ modules/lang-painless/src/main/java/org/elasticsearch/painless/Metadata.java
@@ -20,11 +20,8 @@
 package org.elasticsearch.painless;
 
 import org.antlr.v4.runtime.ParserRuleContext;
-import org.antlr.v4.runtime.tree.ParseTree;
 import org.elasticsearch.painless.Definition.Cast;
 import org.elasticsearch.painless.Definition.Type;
-import org.elasticsearch.painless.PainlessParser.ExpressionContext;
-import org.elasticsearch.painless.PainlessParser.PrecedenceContext;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -37,7 +34,6 @@ import java.util.Map;
  * the root of the ANTLR parse tree, and the {@link CompilerSettings}.
  */
 class Metadata {
-
     /**
      * StatementMetadata is used to store metadata mostly about
      * control flow for ANTLR nodes related to if/else, do, while, for, etc.
@@ -387,15 +383,6 @@ class Metadata {
     }
 
     /**
-     * A utility method to output consistent error messages.
-     * @param ctx The ANTLR node the error occurred in.
-     * @return The error message with tacked on line number and character position.
-     */
-    static String error(final ParserRuleContext ctx) {
-        return "Error [" + ctx.getStart().getLine() + ":" + ctx.getStart().getCharPositionInLine() + "]: ";
-    }
-
-    /**
      * Acts as both the Painless API and white-list for what types and methods are allowed.
      */
     final Definition definition;
@@ -490,7 +477,7 @@ class Metadata {
         final StatementMetadata sourcesmd = statementMetadata.get(source);
 
         if (sourcesmd == null) {
-            throw new IllegalStateException(error(source) + "Statement metadata does not exist at" +
+            throw new IllegalStateException("Statement metadata does not exist at" +
                 " the parse node with text [" + source.getText() + "].");
         }
 
@@ -518,7 +505,7 @@ class Metadata {
         final ExpressionMetadata sourceemd = expressionMetadata.get(source);
 
         if (sourceemd == null) {
-            throw new IllegalStateException(error(source) + "Expression metadata does not exist at" +
+            throw new IllegalStateException("Expression metadata does not exist at" +
                 " the parse node with text [" + source.getText() + "].");
         }
 
@@ -546,7 +533,7 @@ class Metadata {
         final ExternalMetadata sourceemd = externalMetadata.get(source);
 
         if (sourceemd == null) {
-            throw new IllegalStateException(error(source) + "External metadata does not exist at" +
+            throw new IllegalStateException("External metadata does not exist at" +
                 " the parse node with text [" + source.getText() + "].");
         }
 
@@ -574,7 +561,7 @@ class Metadata {
         final ExtNodeMetadata sourceemd = extNodeMetadata.get(source);
 
         if (sourceemd == null) {
-            throw new IllegalStateException(error(source) + "External metadata does not exist at" +
+            throw new IllegalStateException("External metadata does not exist at" +
                 " the parse node with text [" + source.getText() + "].");
         }
